requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
pydantic
aiomysql
tenacity